Perform platform-agnostic configuration centrally
