+++ .kitchen.travis.yml
@@ -12,3 +12,17 @@ suites:
         package_cloud_token: <%= ENV['PACKAGECLOUD_TOKEN'] %>
         package_cloud_repo: snoopy
         publish_artifacts: <%= ENV['TRAVIS_PULL_REQUEST'] == 'false' ? true : false %>
+        build_version: <%=
+          require_relative 'libraries/helpers_coordinator'
+          SnoopyBuildCookbook::Helpers::Coordinator.version
+        %>
+        build_revision: <%=
+          require_relative 'libraries/helpers_coordinator'
+          conf = {
+            user: 'socrata-platform',
+            token: ENV['PACKAGECLOUD_TOKEN'],
+            repo: 'snoopy'
+          }
+          SnoopyBuildCookbook::Helpers::Coordinator.configure!(conf)
+          SnoopyBuildCookbook::Helpers::Coordinator.revision
+        %>
