+++ README.md
@@ -42,14 +42,36 @@ results, and publish the new package to PackageCloud.io.
 Should the need arise, `fpm-cook` commands can still be run manually from
 the recipe directory in `files/default/fpm-recipes/snoopy/`.
 
+Recipes
+=======
+
+***default***
+
+Ties together all the other recipes.
+
+***_configure***
+
+Configures the helper class for use by the other recipes.
+
+***_build***
+
+Runs a package build with FPM-Cookery.
+
+***_verify***
+
+Installs the artifact generated by the `_build` recipe and runs some acceptance
+tests against it with ServerSpec.
+
+***_deploy***
+
+If configured to publish artifacts, uploads the result of the `_build` recipe
+to PackageCloud.
+
 Attributes
 ==========
-The included attributes file controls the build version and revision to use
-for the package, defaulting to the most recent version of snoopy released and a
-build of 1.
-
-    default['snoopy_build']['build_version']
-    default['snoopy_build']['build_revision']
+The included attributes file defines attributes for PackageCloud access, the
+build version and revision number, and a boolean controlling whether or not
+artifacts should be published to PackageCloud when completed.
 
 Contributing
 ============
