+++ recipes/_verify.rb
@@ -20,14 +20,16 @@
 
 gem_package 'serverspec'
 
+artifact_path = SnoopyBuildCookbook::Helpers::Builder.package_file
+
 case node['platform_family']
 when 'debian'
   dpkg_package 'snoopy' do
-    package_name lazy { SnoopyBuildCookbook::Helpers.package_file }
+    package_name artifact_path
   end
 when 'rhel'
   rpm_package 'snoopy' do
-    package_name lazy { SnoopyBuildCookbook::Helpers.package_file }
+    package_name artifact_path
   end
 end
