+++ spec/spec_helper.rb
@@ -9,6 +9,8 @@ require 'coveralls'
 require 'tmpdir'
 require 'fileutils'
 require_relative '../libraries/helpers'
+require_relative '../libraries/helpers_builder'
+require_relative '../libraries/helpers_coordinator'
 
 RSpec.configure do |c|
   c.color = true
